% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculoMediaDpGenericoComPat.R
\name{calculoMediaDpGenericoComPat}
\alias{calculoMediaDpGenericoComPat}
\title{Calculo da media e desvio padrao de tabela Nwlistop COM patamar}
\usage{
calculoMediaDpGenericoComPat(pasta, nomeTabela, passo, colunaInicialJaneiro)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com as tabelas do Nwlistop}

\item{nomeTabela}{nome da tabela do Nwlistop ex (efiol)}

\item{passo}{tamanho do campo}

\item{colunaInicialJaneiro}{coluna em que inicia a impressao dos dados (coluna posterior ao final da impressao do patamar)}
}
\value{
\code{df.cmoMedio} data frame com os valores de custo marginal de demanda
\itemize{
\item codigo do REE (\code{$codREE})
\item valor do ano (\code{$ano})
\item media da variavel (\code{$media})
\item desvio padrao da media da variavel (\code{$dp})
}
}
\description{
Faz o calculo da media e do desvio padrao da variavel de interesse por submercado e por ano de estudo,
}
\examples{
\dontrun{
calculoMediaDpGenericoComPat("C:/PDE2027_Caso080", "ghtot", 9, 12)}

}
