% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidaHidreletricasnoHorizonte.R
\name{consolidaHidreletricasnoHorizonte}
\alias{consolidaHidreletricasnoHorizonte}
\title{Dados consolidados de hidreletricas ao longo do horizonte}
\usage{
consolidaHidreletricasnoHorizonte(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE.}
}
\value{
\code{df.consolidadoUHEnoHorizonte} data frame com dados consolidados de hidreletricas ao longo do horizonte
\itemize{
\item codigo da usina no cadastro de usinas hidroeletricas (\code{$codUsina})
\item nome da usina (\code{$nomeUsina})
\item ano e mes de entrada em operacao da unidade hidroeletrica (\code{$anoMes})
\item codigo do REE (\code{$codREE})
\item canal de fuga medio [m] (\code{$canalFugaMedio})
\item volume Maximo [hm3] (\code{$volumeMaximoFinal})
\item volume Minimo [hm3] (\code{$volumeMinimoFinal})
\item vazao minima do historico [m3/s] (\code{$vazaoMinimaFinal})
\item numero de conjuntos de maquinas (\code{$numeroConjuntos})
\item taxa esperada de indisponibilidade forcada (TEIF) [\%] (\code{$TEIF})
\item indisponibilidade programada (IP) [\%] (\code{$IP})
\item potencia efetiva da usina [MW] (\code{$potenciaEfetivaUsina})
}
}
\description{
Faz a consolidacao dos dados das hidreletricas ao longo do horizonte de estudo definido no NEWAVE.
Usa como referencia para a leitura dos arquivos as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
consolidaHidreletricasnoHorizonte(pastaCaso)
}

}
