% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definePeriodo.R
\name{definePeriodo}
\alias{definePeriodo}
\title{Data frama com periodo do caso NEWAVE}
\usage{
definePeriodo(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{df.periodo} data frame anos e meses do horizonte do estudo
\itemize{
\item ano (\code{$ano})
\item mes (\code{$mes})
\item anoMes (\code{$anoMes})
}
}
\description{
Com o auxilio da funcao \code{\link{leituraDadosGerais}} cria um data frame com o horizonte do caso
}
\examples{
\dontrun{
definePeriodo("C:/PDE2027_Caso080")
}

}
