% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escreveHidr.R
\name{escreveHidr}
\alias{escreveHidr}
\title{Escritor de dados das usinas hidroeletricas}
\usage{
escreveHidr(lt.dadosUsinasHidroeletricas, arquivo)
}
\arguments{
\item{lt.dadosUsinasHidroeletricas}{lista com os dados de usinas 
hidroeletricas no mesmo formato obtido na leitura}

\item{arquivo}{caracter com o nome do arquivo que sera criado}
}
\description{
Faz a escrita do arquivo do NEWAVE com dados das usinas hidroeletricas 
(hidr.dat).
Arquivo de acesso direto, nao formatado não formatado, com 320 / 600 
postos, cada registro correspondendo a uma usina.
}
\examples{
\dontrun{
escreveHidr(lt.dadosUsinasHidroeletricas, "C:/PDE2027_Caso080/hidr.dat")
}

}
