% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escreveVazoes.R
\name{escreveVazoes}
\alias{escreveVazoes}
\title{Escritor de dados de vazoes historicas}
\usage{
escreveVazoes(df.vazoes, arquivo)
}
\arguments{
\item{df.vazoes}{data frame com os dados de vazoes naturais historicas no
mesmo formato obtido na leitura}

\item{arquivo}{caracter com o nome do arquivo que sera criado}
}
\description{
Faz a escrita do arquivo do NEWAVE com dados de vazoes naturais historicas 
afluentes as usinas hidroeletricas (vazoes.dat).
Arquivo de acesso direto, nao formatado não formatado, com 320 / 600 
postos, cada registro correspondendo a um mes do historico.
}
\examples{
\dontrun{
escreveVazoes(df.vazoes, "C:/PDE2027_Caso080/vazoes.dat")
}

}
