% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraArquivos.R
\name{leituraArquivos}
\alias{leituraArquivos}
\title{Leitor dos nomes dos arquivos do NEWAVE}
\usage{
leituraArquivos(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao do arquivo caso.dat.}
}
\value{
\code{df.arquivos} data frame com os arquivos utilizados pelo NEWAVE.
\itemize{
\item descricao (\code{$descricao})
\item nomes dos arquivos (\code{$arquivo})
}
}
\description{
Procura o arquivo caso.dat (nos formatos Caso.dat, caso.dat e CASO.DAT) e faz a leitura do arquivo do CEPEL
indicado nele (geralmente arquivos.dat) para recuperar a lista de arquivos de entrada e resultados utilizados pelo NEWAVE.
}
\examples{
\dontrun{
leituraArquivos("C:/PDE2027_Caso080")}

}
