% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraArquivosDC.R
\name{leituraArquivosDC}
\alias{leituraArquivosDC}
\title{Leitor dos nomes dos arquivos do DECOMP}
\usage{
leituraArquivosDC(pastaCaso)
}
\arguments{
\item{pastaCaso}{string com localizacao do arquivo caso.dat.}
}
\value{
\code{df.arquivos} data frame com os arquivos utilizados pelo DECOMP.
\itemize{
\item nomes dos arquivos (\code{$arquivo})
}
}
\description{
Procura o arquivo caso.dat (nos formatos Caso.dat, caso.dat e CASO.DAT) e faz a leitura do arquivo do CEPEL
indicado nele (geralmente rvX) para recuperar a lista de arquivos de entrada utilizados pelo DECOMP.
}
\examples{
\dontrun{
leituraArquivosDC("C:/Pasta_Deck_DECOMP")
}

}
