% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraClassesTermicas.R
\name{leituraClassesTermicas}
\alias{leituraClassesTermicas}
\title{Leitor dos dados de classes termicas}
\usage{
leituraClassesTermicas(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE.}
}
\value{
\code{df.ClassesTermicas} data frame com dados de usinas termoeletricas
\itemize{
\item numero da usina termica (\code{$codUsinaTermica})
\item periodo (anoMes-AAAAMM) (\code{$anoMes})
\item tipo de combustivel da classe termica (\code{$tipoComb})
\item custo de operacao da classe termica ($/MWh) (\code{$CVU})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados de classes termicas (clast.*).
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraClassesTermicas("C:/PDE2027_Caso080")}

}
