% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraCustoMarginalDemanda.R
\name{leituraCustoMarginalDemanda}
\alias{leituraCustoMarginalDemanda}
\title{Leitor de dados de custo marginal de demanda do submercado}
\usage{
leituraCustoMarginalDemanda(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de custo marginal de demanda}
}
\value{
\code{df.custoMarginalDemanda} data frame com os valores de custo marginal de demanda
\itemize{
\item codigo do submercado (\code{$codSubmercado})
\item serie (\code{$serie})
\item numero do patamar de mercado (\code{$patamar})
\item valor de ano e mes (\code{$anoMes})
\item valor do custo marginal de demanda [$/MWh] (\code{$custoMarginalDemanda})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de custo marginal de demanda do submercado  (cmargxxx.*) e 
recupera esses valores por ano, mes, patamar e serie.
Nao retorna os valores de total, media, desvio e etc. do arquivo de origem. 
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas. 
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraCustoMarginalDemanda("C:/PDE2027_Caso080")}

}
