% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraDadosDuracaoPatamar.R
\name{leituraDadosDuracaoPatamar}
\alias{leituraDadosDuracaoPatamar}
\title{Leitor dos dados de duracao de patamares de mercado}
\usage{
leituraDadosDuracaoPatamar(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE.}
}
\value{
\code{df.dadosDuracaoPatamar} data frame com dados de patamares de carga
\itemize{
\item numero do patamar (\code{$patamar})
\item ano e mes do dado (\code{$anoMes})
\item duracao do patamar (\code{$duracaoPatamar})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados de duracao de patamares (patamar.*).
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraDadosDuracaoPatamar("C:/PDE2027_Caso080")
}

}
