% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraDadosExpansaoUsinasHidro.R
\name{leituraDadosExpansaoUsinasHidro}
\alias{leituraDadosExpansaoUsinasHidro}
\title{Leitor de dados de expansao de usinas hidroeletricas}
\usage{
leituraDadosExpansaoUsinasHidro(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{lt.dadosExpansaoHidro} lista com data frames com dados de expansao hidroeletrica
\itemize{
\item \code{df.dadosVolumeMorto} data frame com dados de enchimento de volume morto
\itemize{
\item codigo da usina no cadastro de usinas hidroeletricas (\code{$codUsina})
\item nome da usina (\code{$nomeUsina})
\item ano e mes de inicio do enchimento de volume morto (\code{$anoMes})
\item duracao em meses do enchimento de volume morto (\code{$duracaoEnchimentoVM})
\item \% do volume morto ja preenchido ate a data de inicio informada (\code{$volumeMortoCheio})
}
\item \code{df.dadosExpansaoHidro} data frame com dados de expansao hidroeletrica no "mesmo formato" do original
\itemize{
\item codigo da usina no cadastro de usinas hidroeletricas (\code{$codUsina})
\item nome da usina (\code{$nomeUsina})
\item conjunto de maquinas (\code{$conjunto})
\item ano e mes de entrada em operacao da unidade hidroeletrica (\code{$anoMes})
\item potencia unitaria da unidade adicionada (\code{$potenciaUnitaria})
\item numero da unidade a ser adicionada (\code{$unidadeAdicionada})
}
\item \code{df.dadosExpansaoHidroTempo} data frame com dados de expansao hidroeletrica no tempo
\itemize{
\item codigo da usina no cadastro de usinas hidroeletricas (\code{$codUsina})
\item nome da usina (\code{$nomeUsina})
\item conjunto de maquinas (\code{$conjunto})
\item ano e mes de entrada em operacao da unidade hidroeletrica (\code{$anoMes})
\item potencia unitaria da unidade adicionada (\code{$potenciaUnitaria})
\item numero de maquinas (\code{$numeroMaquinas})
}
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados de expansao de usinas hidroeletricas (exph.*). Usa as funcoes internas \code{\link{leituraArquivos}},
\code{\link{definePeriodo}}.
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
Retorna as ateracoes no tempo (por todos os meses e anos do horizonte mesmo que o valor seja o mesmo) e sem os periodos de pre e pos
}
\examples{
\dontrun{
leituraDadosExpansaoUsinasHidro("C:/PDE2027_Caso080")
}

}
