% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraDadosGerais.R
\name{leituraDadosGerais}
\alias{leituraDadosGerais}
\title{Leitor dos dados gerais}
\usage{
leituraDadosGerais(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{df.dadosGerais} data frame com informacao sobre o estudo
\itemize{
\item tipo de execucao (0 so executa simulacao final; 1 rodada completa) (\code{$tipoExecucao})
\item duracao de cada estagio de operacao em meses  (\code{$duracaoPeriodo})
\item numero de anos de duracao do estudo (\code{$duracaoEstudo})
\item mes inicial do periodo que antecede o periodo de planejamento (\code{$mesInicioPre})
\item mes do inicio do estudo (\code{$mesInicio})
\item ano do inicio do estudo (\code{$anoInicio})
\item numero de anos iniciais para fins de estabilizacao no calculo da politica (\code{$anosPre})
\item numero de anos finais para fins de estabilizacao no calculo da politica (\code{$anosPos})
\item numero de anos finais para fins de estabilizacao na simulacao final (\code{$anosPosFinal})
\item controle de impressao das carcteristicas das usinas (0 nao imprime; 1 imprime) (\code{$dados})
\item controle de impressao dos dados de mercado de energia (0 nao imprime; 1 imprime) (\code{$mercados})
\item controle de impressao das energias hitoricas afluentes (0 nao imprime; 1 imprime) (\code{$energias})
\item controle de impressao dos parametros do modelo estocastico (0 nao imprime; 1 imprime) (\code{$modeloEstocastico}) 
\item controle de impressao dos parametros dos reservatorios equivalentes de energia (0 nao imprime; 1 imprime) (\code{$subsistema})
\item numero maximo de iteracoes (\code{$maximoIteracoes})
\item numero de simulacoes forward (\code{$simulacoesForward})
\item numero de aberturas para a simulacao backward (\code{$aberturasSimulacaoBackward}) 
\item numero de aberturas variavel por periodo (0 nao sera considerado; 1 sera considerado) (\code{$aberturasVariavel})
\item numero de series sinteticas (\code{$seriesSinteticas})
\item ordem maxima do modelo estocastico PAR(p) (\code{$ordemMaximaParP})
\item ano inicial do arquivo de vazoes historicas (\code{$anoInicioHistorica})
\item tamanho do registro do arquivo de vazoes historicas (0 para 320 palavras; 1 para 600 palavras) (\code{$tamanhoVazoesHistorica})
\item calculo de energia armazenada inicial (0 usa o valor do volume inicial deste arquivo; 1 usa valor do arquivo de configuracao hidroeletrica) (\code{$calculaVolumeInicial})
\item probabilidade associada ao intervalo de confianca para convergencia do algoritmo (%) (\code{$tolerancia})
\item taxa de desconto anual (\code{$taxaDesconto})
\item simulacao final apos convergencia PDDE (0 nao simula; 1 series sinteticas; 2 serie historica; 3 consistencia de dados) (\code{$tipoSimulacao})
\item controle de impressao dos resultados da simulacao final e do calculo da politica (0 sinopse; 1 op. detalhada) (\code{$impressaoSimulacao}) 
\item controle de impressao dos riscos de deficit e valor esperado da ENS (0 convergencia final apenas; 1 total as iteracoes) (\code{$impressaoConvergencia})
\item de quantas em quantas series sera gravado o relatorio detalhado da simulacao final (\code{$intervaloGravar})
\item numero minimo de iteracoes para a convergencia da politica (\code{$minimoIteracoes})
\item indica a iteracao partir da qual sera investigada a incerteza do parametro ZINF (\code{$iteracaoApartir})
\item indica a adocao ou nao de corte de carga por otimizacao (CCO) na simulacao final (0 nao considera na simulacao final; 1 considera) (\code{$racionamento})
\item numero de anos de informacoes de manutencao programada de usinas termicas a serem considerados no arquivo de dados de manutencao termica (item 0) (\code{$anosManutencao})
\item consideracao da tendencia hidrologica no calculo da politica (0 nao sera lido; 1 sera lido por REE; 2 sera lido por posto de medicao) (\code{$tendenciaHidroPolitica})
\item consideracao da tendencia hidrologica na simulacao final (0 nao sera lido; 1 sera lido por REE; 2 sera lido por posto de medicao) (\code{$tendenciaHidroSimFinal})
\item consideracao das restricoes de itaipu (0 nao considera; 1 considera) (\code{$restricaoItaipu})
\item consideracao do bid de demanda (0 nao considera; 1 considera) (\code{$bidDemanda})
\item consideracao das perdas de transmissao (0 nao considera; 1 considera) (\code{$perdasTransmissao})
\item consideracao do el nino (0 nao considera; 1 considera) (\code{$elNino})
\item indice de identificao ENSO (\code{$indiceEnso})
\item tipo de duracao do patamar (0 sazonal; 1 variavel por ano) (\code{$duracaoPatamar})
\item consideracao desvio da agua (0 nao considera; 1 considera) (\code{$desvioAgua})
\item consideracao da energia de desvio de agua como funcao da energia armazenada (0 constante; 1 variavel com o armazenamento) (\code{$correcaoDesvio})
\item controle da curva de seguranca (0 sem car d VMINP; 1 car e/ou VMINP) (\code{$controleCurvaSeguranca})
\item controle da geracao de cenario de afluencias para as simulacoes backward e forward (0 ruidos fw sorteados da bw e compensacao correlacao espacial; 1 compensacao bw; 2 compensacao bw e fw) (\code{$tipoGeracao})
\item profundidade para calculo do risco de deficit (%) - primeiro valor (\code{$riscoDeficitPrimeiro})
\item profundidade para calculo do risco de deficit (%) - segundo valor (\code{$riscoDeficitSegundo}) 
\item funcionalidade pseudo-partida quente: numero de iteracoes a ser considerada para a simulacao final (0 considera todas as iteracoes) (\code{$pseudoPartidaQuente})
\item consideracao de agrupamento livre de intercambios (0 nao considera; 1 considera) (\code{$agrupamentoIntercambio})
\item consideracao de equalizacao de penalidades de intercambio (\code{$equalizacaoIntercambio}) 
\item consideracao da representacao da submortizacao (0 nao considera; 1 subsistema; 2 usina) (\code{$submortizacao})
\item consideracao da ordenacao automatica de subsistemas/submercados e classes termicas (0 nao considera; 1 considera) (\code{$ordenacaoAutomatica})
\item consideracao do arquivo de cargas adicionais (0 nao considera; 1 considera) (\code{$cargaAdicional})
\item valor percentual de ZSUP a ser subtraido de LINF para o criterio de parada estatistico (%) (\code{$deltaZsup}) 
\item valor maximo percentual para delta de ZINF no criterio de parada nao estatistico (%) (\code{$deltaZinf})  
\item numero de deltas de ZINF consecutivos a serem considerados no criterio nao estatistico (\code{$iteracoesDeltaZinf}) 
\item consideracao de despacho antecipado de usinas termicas a gas natural liquefeito (GNL) (0 nao considera; 1 considera) (\code{$despachoAntecipado})
\item modificacao automatica do montante de antecipacao de despacho de uma usina GNL quando a capacidade de geracao maxima desta usina for inferior a este valor (0 nao considera; 1 considera) (\code{$modificacaoAuto})
\item consideracao de restricoes de geracao hidraulica minima (0 nao considera; 1 considera) (\code{$geracaoHidraulica}) 
\item mes de inicio para o calculo da simulacao final (\code{$mesSimulacao}) 
\item ano de inicio para o calculo da simulacao final (\code{$anoSimulacao})
\item volume armazenado inicial (%) por REE para calculo da simulacao final (\code{$volumeSimulacao})
\item utilizacao do gerenciador externo de processos (0 nao considera; 1 considera) (\code{$gerenciadorExterno}) 
\item utilizacao da comunicacao em dois niveis (0 nao considera; 1 considera) (\code{$comunicaDoisNiveis}) 
\item utilizacao de armazenamento local de arquivos temporarios (0 nao considera; 1 considera) (\code{$arquivoTemporario})
\item utilizacao de alocacao em memoria da energia natural afluente (0 nao considera; 1 considera) (\code{$energiaAfluente})
\item utilizacao de alocacao em memoria dos cortes da funcao de custo futuro (0 nao considera; 1 considera) (\code{$cortesCustoFuturo})
\item utilizacao de mecanismo de aversao a risco: SAR (0 nao considera; 1 considera) (\code{$riscoSar})
\item utilizacao de mecanismo de aversao a risco: CVaR (0 nao considera; 1 considera constante; 2 considera variavel) (\code{$riscoCvar})
\item consideracao do criterio de minimo ZSUP para convergencia (0 nao considera; 1 considera) (\code{$zsup}) 
\item consideracao do requisito de vazao minima (0 nao considera; 1 considera) (\code{$vazaoMinima}) 
\item consideracao de restricoes eletricas internas aos REEs (0 nao considera; 1 considera) (\code{$restricaoEletrica})
\item consideracao do procedimento de Selecao de Cortes de Benders (0 nao considera; 1 considera) (\code{$selecaoCortes})
\item consideracao de Janela de Cortes de Benders (0 nao considera; 1 considera) (\code{$janelaCortes})
\item consideracao Reamostragem de Cenarios (0 nao considera; 1 considera) (\code{$reamostragem})
\item tipo de reamostragem de cenarios (0 recombinacao; 1 plena) (\code{$tipoReamostragem})
\item passo para reamostragem de cenarios (permitido valores entre 1 e numero maximo de iteracoes) (\code{$passoReamostragem})
\item consideracao do No Zero no calculo de ZINF (0 nao considera; 1 considera) (\code{$noZero})
\item consulta a funcao de custo futuro ao longo das iteracoes da PDDE (0 nao considera; 1 considera) (\code{$custoFuturo})
\item impressao dos cenarios de ENA (0 nao imprime; 1 imprime) (\code{$cenariosEna})
\item impressao dos cortes ativos (0 nao imprime; 1 imprime) (\code{$cortesAtivo})
\item escolha do representante do processo agregacao (0 mais proximo; 1 centroide) (\code{$representanteAgregacao})
\item escolha da matriz de correlacao espacial (0 anual; 1 mensal) (\code{$correlacaoEspacial})
\item desconsidera criterio estatistico no processo de convergencia (0 nao; 1 sim) (\code{$convergencia})
\item momento de realizacao da reamostragem (0 backward; 1 forward) (\code{$momentoReamostragem})
\item manutencao dos arquivos de ENA gerados para reamostragem apos a execucao do programa (0 apaga; 1 mantem) (\code{$arquivosEna})
\item teste da convergencia somente apos iteracao minima (inclusive) (0 primeira iteracao; 1 iteracao minima) (\code{$iteracaoMinima})

}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados gerais (dger.*) e recupera seus dados. Usa a funcao \code{\link{leituraArquivos}}.
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraDadosGerais("C:/PDE2027_Caso080")}

}
