% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraDadosProfundidadePatamarCarga.R
\name{leituraDadosProfundidadePatamarCarga}
\alias{leituraDadosProfundidadePatamarCarga}
\title{Leitor dos dados de profundidade patamares de carga}
\usage{
leituraDadosProfundidadePatamarCarga(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE.}
}
\value{
\code{df.dadosProfundidadePatamarCarga} data frame com dados de profundidade de patamares de carga
\itemize{
\item sistema (\code{$codSubsistema})
\item numero do patamar (\code{$patamar})
\item ano e mes do dado (\code{$anoMes})
\item profundidade de patamar de carga (\code{$profundidadeCarga})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados de profundidade de patamares de carga (patamar.*).
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraDadosProfundidadePatamarCarga("C:/PDE2027_Caso080")
}

}
