% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraDadosSubsistema.R
\name{leituraDadosSubsistema}
\alias{leituraDadosSubsistema}
\title{Leitor dos dados dos subsistemas/submercados}
\usage{
leituraDadosSubsistema(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{df.subsistema} data frame com os dados dos subsistemas/submercados
\itemize{
\item numero do subsistema/submercado (\code{$codSubsistema}) 
\item nome do subsistema/submercado (\code{$nomeSubsistema})
\item tipo de subsistema: ficticio [1], ou real [0] (\code{$tipoFicticio})
\item custo de deficit em $/MWh (\code{$deficit})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados dos subsistemas/submercados (sistema.*) 
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraSubsistema("C:/PDE2027_Caso080") }

}
