% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraDeficitEnergia.R
\name{leituraDeficitEnergia}
\alias{leituraDeficitEnergia}
\title{Leitor de dados de deficit de energia}
\usage{
leituraDeficitEnergia(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de deficit de energia}
}
\value{
\code{df.deficitEnergia} data frame com os valores de deficit de energia
\itemize{
\item codigo do submercado (\code{$codSubmercado})
\item numero do patamar com deficit (\code{$patamarDeficit})
\item serie (\code{$serie})
\item numero do patamar de mercado (\code{$patamar})
\item valor de ano e mes (\code{$anoMes})
\item valor do deficit de energia [MWmes] (\code{$deficitEnergia})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de deficit de energia por subsistema e por deficit no patamar  (defXXXpXXX.*) e
recupera esses valores por defcit, ano, mes, patamar e serie.
Nao retorna os valores de total, media, desvio e etc. do arquivo de origem.
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas.
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraDeficitEnergia("C:/PDE2027_Caso080")
}

}
