% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraDeficitEnergiaSin.R
\name{leituraDeficitEnergiaSin}
\alias{leituraDeficitEnergiaSin}
\title{Leitor de dados de deficit de energia (Sistema Interligado Nacional)}
\usage{
leituraDeficitEnergiaSin(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de deficit de energia do sistema interligado}
}
\value{
\code{df.deficitEnergiaSin} data frame com os valores de deficit de energia do sistema interligado
\itemize{
\item numero do patamar com deficit (\code{$patamarDeficit})
\item serie (\code{$serie})
\item numero do patamar de mercado (\code{$patamar})
\item valor de ano e mes (\code{$anoMes})
\item valor do deficit de energia [MWmes] (\code{$deficitEnergia})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de deficit de energia por deficit no patamar do sistema interligado nacional   (defsinpXXX.*) e
recupera esses valores por patamar com defcit, ano, mes, patamar e serie.
Nao retorna os valores de total, media, desvio e etc. do arquivo de origem.
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas.
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraDeficitEnergiaSin("C:/PDE2027_Caso080")
}

}
