% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraDeficitSistema.R
\name{leituraDeficitSistema}
\alias{leituraDeficitSistema}
\title{Leitor dos dados de deficit de sistema}
\usage{
leituraDeficitSistema(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{df.sistema} data frame com informacoes de sistema
\itemize{
  \item codigo do subsistema/submercado (\code{$codSubsistema})
  \item nome do subsistema/submercado (\code{$nomeSubsistema})
  \item tipo de subsistema/submercado (\code{$tipoFicticio}) - pode assumir 1 para ficticio, ou 0 para real
  \item patamar de carga (\code{$patamar})
  \item custo de deficit em R$/MWh (\code{$custoDeficit})
  \item profundidade de deficit em p.u. (\code{$profundidadeDefict})
  }
}
\description{
Faz a leitura do arquivo do NEWAVE com informacao de deficit do sistema (sistema.d**). Usa a funcao \code{\link{leituraArquivos}}
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraDeficitSistema("C:/PDE2027_Caso080")}

}
