% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraEditorHidr.R
\name{leituraEditorHidr}
\alias{leituraEditorHidr}
\title{Leitor de dados das planilha Editor do HIDR}
\usage{
leituraEditorHidr(planilha, nRegistros = 600, tamanhoRegistrosPoli = 4)
}
\arguments{
\item{planilha}{caracter com localizacao da planilha do Editor do HIDR}

\item{nRegistros}{numero de registros do arquivo hidr (320/600)}

\item{tamanhoRegistrosPoli}{tamanho dos registros dos dados de polinomio em
bytes 4/8}
}
\description{
Faz a leitura do arquivo do .xlsm utilizado na EPE com dados das usinas 
hidroeletricas no formato utilizado no pacote leitorrmpe.
}
\examples{
\dontrun{
leituraEditorHidr("C:/PDE2027_Caso080/Editor do HIDR.xlsm", 600, 4)
}

}
