% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraEnergiaArmazenadaFinal.R
\name{leituraEnergiaArmazenadaFinal}
\alias{leituraEnergiaArmazenadaFinal}
\title{Leitor dos dados de energia armazenada final por REE}
\usage{
leituraEnergiaArmazenadaFinal(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de energia armazenada final}
}
\value{
\code{df.energiaArmazenadaFinal} data frame com os valores de energia armazenada final por REE
\itemize{
\item codigo do REE (\code{$codREE})
\item serie (\code{$serie})
\item valor de ano e mes (\code{$anoMes})
\item valor de energia armazenada final [MWMes] (\code{$earmf})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com informacao de energia armazenada final por REE (earmfXXX.out) e recupera esses valores por ano, mes e serie.
Nao retorna os valores de media, desvio e etc. do arquivo de origem.
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas. 
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraEnergiaArmazenadaFinal("C:/PDE2027_Caso080")}

}
