% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraEnergiaArmazenadaFinalPercentualSin.R
\name{leituraEnergiaArmazenadaFinalPercentualSin}
\alias{leituraEnergiaArmazenadaFinalPercentualSin}
\title{Leitor de dados de armazenamento percentual para o Sistema Interligado Nacional}
\usage{
leituraEnergiaArmazenadaFinalPercentualSin(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de armazenamento percentual do SIN}
}
\value{
\code{df.earmfpSin} data frame com os valores de energia vertida para o SIN
\itemize{
\item serie (\code{$serie})
\item valor de ano e mes (\code{$anoMes})
\item valor de armazenamento percentual [%] (\code{$earmfp})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de armazenamento percentual para o Sistema Interligado Nacional  (earmfpsin.*) e recupera esses valores por ano, mes e serie.
Nao retorna os valores de media, desvio e etc. do arquivo de origem.
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas.
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraEnergiaVertidaSin("C:/PDE2027_Caso080")
}

}
