% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraEnergiaArmazenadaFinalSubmercado.R
\name{leituraEnergiaArmazenadaFinalSubmercado}
\alias{leituraEnergiaArmazenadaFinalSubmercado}
\title{Leitor dos dados de energia armazenada final por Submercado}
\usage{
leituraEnergiaArmazenadaFinalSubmercado(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de energia armazenada final por submercado}
}
\value{
\code{df.energiArmazenadaFinalSubmercadoSubmercado} data frame com os valores de energia armazenada final por REE
\itemize{
\item codigo do submercado (\code{$codSubmercado})
\item serie (\code{$serie})
\item valor de ano e mes (\code{$anoMes})
\item valor de energia armazenada final [MWMes] (\code{$earmf})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com informacao de energia armazenada final por Submercado (earmfMXXX.out) e recupera esses valores por ano, mes e serie.
Nao retorna os valores de media, desvio e etc. do arquivo de origem.
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas. 
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraEnergiArmazenadaFinalSubmercado("C:/PDE2027_Caso080")}

}
