% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraGeracaoHidroTotal.R
\name{leituraGeracaoHidroTotal}
\alias{leituraGeracaoHidroTotal}
\title{Leitor de dados de geracao hidroeletrica total}
\usage{
leituraGeracaoHidroTotal(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de geracao hidroeletrica total}
}
\value{
\code{df.geracaoHidroTotal} data frame com os valores de geracao hidroeletrica total por REE
\itemize{
\item codigo do REE (\code{$codREE})
\item serie (\code{$serie})
\item numero do patamar de mercado (\code{$patamar})
\item valor de ano e mes (\code{$anoMes})
\item valor de geracao hidroeletrica [MWmes] (\code{$geracao})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de geracao hidroeletrica total por REE (ghtotxxx.*) e recupera esses valores por ano, mes, patamar e serie.
Nao retorna os valores de total, media, desvio e etc. do arquivo de origem. 
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas. 
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraGeracaoHidroTotal("C:/PDE2027_Caso080")}

}
