% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraGeracaoHidroUsina.R
\name{leituraGeracaoHidroUsina}
\alias{leituraGeracaoHidroUsina}
\title{Leitor de dados de geracao hidro por usina}
\usage{
leituraGeracaoHidroUsina(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de geracao hidro por usina}
}
\value{
\code{df.geracaoHidroUsina} data frame com os valores de geracao hidro por usina
\itemize{
\item codigo da usina (\code{$codUsina})
\item serie (\code{$serie})
\item numero do patamar de mercado (\code{$patamar})
\item valor de ano e mes (\code{$anoMes})
\item valor da geracao hidro [MWmes] (\code{$geracao})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de geracao hidro por usina  (ghiduhxxx.*) e
recupera esses valores por codigo da usina, ano, mes, patamar e serie.
Nao retorna os valores de total, media, desvio e etc. do arquivo de origem.
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas.
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraGeracaoHidroUsina("C:/PDE2027_Caso080")
}

}
