% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraGeracaoTermicaClassesTotal.R
\name{leituraGeracaoTermicaClassesTotal}
\alias{leituraGeracaoTermicaClassesTotal}
\title{Leitor de dados de geracao termica por classes e total}
\usage{
leituraGeracaoTermicaClassesTotal(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de geracao termica por classes e total}
}
\value{
\code{df.geracaoTermicaClassesTotal} data frame com os valores de geracao termica por classes e total
\itemize{
\item codigo da usina (\code{$codUsina})
\item codigo do submercado (\code{$codSubmercado})
\item serie (\code{$serie})
\item numero do patamar de mercado (\code{$patamar})
\item valor de ano e mes (\code{$anoMes})
\item valor da geracao termica [MWmes] (\code{$geracao})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de geracao termica por classes e total do submercado  (gtertxxx.*) e
recupera esses valores por codigo da usina, ano, mes, patamar e serie.
Nao retorna os valores de total, media, desvio e etc. do arquivo de origem.
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas.
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraGeracaoTermicaClassesTotal("C:/PDE2027_Caso080")
}

}
