% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraGeracaoTermicaMediaClasses.R
\name{leituraGeracaoTermicaMediaClasses}
\alias{leituraGeracaoTermicaMediaClasses}
\title{Leitor de dados de geracao termica media por classes}
\usage{
leituraGeracaoTermicaMediaClasses(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de geracao termica por classes e total}
}
\value{
\code{df.geracaoTermicaMediaClasses} data frame com os valores de geracao termica media por classes
\itemize{
\item codigo da usina (\code{$codUsina})
\item codigo do submercado (\code{$codSubmercado})
\item valor de ano e mes (\code{$anoMes})
\item valor da geracao termica media [MWmes] (\code{$geracao})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de geracao termica por classes e total do submercado  (gtertxxx.*) e
recupera os valores medios por codigo da usina, ano e mes, na media de todas as series da politica.
}
\examples{
\dontrun{
leituraGeracaoTermicaMediaClasses("C:/PDE2027_Caso080")
}

}
