% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraGeracaoTermicaTotalSin.R
\name{leituraGeracaoTermicaTotalSin}
\alias{leituraGeracaoTermicaTotalSin}
\title{Leitor de dados de geracao termica total (Sistema Interligado Nacional)}
\usage{
leituraGeracaoTermicaTotalSin(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de geracao termica total do sistema interligado nacional}
}
\value{
\code{df.geracaoTermicaTotalSin} data frame com os valores de geracao termica total - SIN
\itemize{
\item serie (\code{$serie})
\item numero do patamar de mercado (\code{$patamar})
\item valor de ano e mes (\code{$anoMes})
\item valor de geracao termica [MWmes] (\code{$geracao})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de geracao termica total do sistema interligado nacional (gttotsin.*) e recupera esses valores por ano, mes, patamar e serie.
Nao retorna os valores de total, media, desvio e etc. do arquivo de origem. 
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas. 
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraGeracaoTermicaTotalSin("C:/PDE2027_Caso080")}

}
