% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraHLiqUsina.R
\name{leituraHLiqUsina}
\alias{leituraHLiqUsina}
\title{Leitor de dados de altura queda liquida por usina}
\usage{
leituraHLiqUsina(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de altura queda liquida por usina}
}
\value{
\code{df.HLiqUsina} data frame com os valores de altura queda liquida por usina
\itemize{
\item codigo da usina (\code{$codUsina})
\item serie (\code{$serie})
\item numero do patamar de mercado (\code{$patamar})
\item valor de ano e mes (\code{$anoMes})
\item valor da altura de queda liquida [m] (\code{$hliq})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de altura de queda liquida por usina  (hliqxxx.*) e
recupera esses valores por codigo da usina, ano, mes, patamar e serie.
Nao retorna os valores de total, media, desvio e etc. do arquivo de origem.
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas.
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraHLiqUsina("C:/PDE2027_Caso080")
}

}
