% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraLimiteAgrupamentoInterligacoes.R
\name{leituraLimiteAgrupamentoInterligacoes}
\alias{leituraLimiteAgrupamentoInterligacoes}
\title{Leitor dos dados do limite de agrupamento de intercambio}
\usage{
leituraLimiteAgrupamentoInterligacoes(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE.}
}
\value{
\code{df.limiteAgrupamentoInterligacoes} data frame com dados de agrupamento livre de interligacoes
\itemize{
\item numero do agrupamento (\code{$codAgrup})
\item ano e mes (\code{$anoMes})
\item patamar (\code{$patamar})
\item limite de agrupamento de intercambio (\code{$limiteAgrup})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados do limite de agrupamento de intercambio (agrint.*).
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraLimiteAgrupamentoInterligacoes("C:/PDE2027_Caso080")
}

}
