% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraLimiteInterligacoes.R
\name{leituraLimiteInterligacoes}
\alias{leituraLimiteInterligacoes}
\title{Leitor dos dados do limite de interligacoes}
\usage{
leituraLimiteInterligacoes(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE.}
}
\value{
\code{df.limiteInterligacoes} data frame com dados de limite de interligacoes
\itemize{
\item subsistema/submercado de origem da interligacao (\code{$codSubsistemaOrigem})
\item subsistema/submercado de destino da interligacao (\code{$codSubsistemaDestino})
\item ano e mes (\code{$anoMes})
\item limite da interligacao (\code{$limiteInterligacao})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados do limite de interligacoes (sistema.*).
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraLimiteInterligacoes("C:/PDE2027_Caso080")
}

}
