% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraManutencoesProgramadas.R
\name{leituraManutencoesProgramadas}
\alias{leituraManutencoesProgramadas}
\title{Leitor dos dados de manutencoes programadas de usinas termetricas}
\usage{
leituraManutencoesProgramadas(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE.}
}
\value{
\code{df.dadosManutencao} data frame com dados de usinas termoeletricas
\itemize{
\item numero da usina termica (\code{$codUsinaTermica})
\item periodo (anoMes-AAAAMM) (\code{$anoMes})
\item indisponibilidade programada (IP) da usina termica (%) (\code{$IP})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados de manutencoes programadas de usinas termetricas (manutt.*).
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraManutencoesProgramadas("C:/PDE2027_Caso80")
}

}
