% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraMercadoEnergia.R
\name{leituraMercadoEnergia}
\alias{leituraMercadoEnergia}
\title{Leitor dos dados de mercado de energia}
\usage{
leituraMercadoEnergia(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE.}
}
\value{
\code{df.mercadoEnergia} data frame com dados de mercado de energia
\itemize{
\item numero do subsistema/submercado (\code{$codSubsistema})
\item ano e mes (\code{$anoMes})
\item mercado de Energia do subsistema/submercado (\code{$energiaMercado})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados de mercado de energia de um subsistema/submercado (sistema.*).
Se houver cargas adicionais (c_adic.*), serao somadas ao subsistema/submercado correspondente.
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraMercadoEnergia("C:/PDE2027_Caso080")
}

}
