% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraMercadoEnergiaPatamar.R
\name{leituraMercadoEnergiaPatamar}
\alias{leituraMercadoEnergiaPatamar}
\title{Leitor dos dados de mercado de energia por patamar}
\usage{
leituraMercadoEnergiaPatamar(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE.}
}
\value{
\code{df.mercadoEnergiaPatamar} data frame com dados de mercado de energia por patamar
\itemize{
\item numero do subsistema/submercado (\code{$codSubsistema})
\item patamar de carga (\code{$patamar})
\item ano e mes (\code{$anoMes})
\item mercado de Energia do subsistema/submercado (\code{$energiaMercado})
}
}
\description{
Utiliza os leitores de mercado de energia e profundidade de patamar de carga para criar um dataframe com os dados de mercado por patamar.
}
\examples{
\dontrun{
leituraMercadoEnergiaPatamar("C:/PDE2027_Caso080")
}

}
