% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraNumPat.R
\name{leituraNumPat}
\alias{leituraNumPat}
\title{Leitor do número de patamares de carga de um caso NEWAVE}
\usage{
leituraNumPat(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE.}
}
\value{
\code{nPat} número de patamares de carga
}
\description{
Faz a leitura do arquivo do NEWAVE com dados de patamares de carga (patamar.*).
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraNumPat("C:/PDE2027_Caso080")
}

}
