% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraNwlistcfCortes.R
\name{leituraNwlistcfCortes}
\alias{leituraNwlistcfCortes}
\title{Leitor NWLISTCF Cortes}
\usage{
leituraNwlistcfCortes(arquivo)
}
\arguments{
\item{arquivo}{caracter indicando o nome do arquivo do NWLISTCF incluindo o caminho. Ex: "C:/nwlistcf.rel"}
}
\value{
\code{df.nwlistcfCortes} data frame com saida do arquivo NWLISTCF
}
\description{
Faz a leitura do arquivo de saida do NEWAVE nwlistcf que gera um relatorio com dados da funcao de custo futuro.
Usa como referencia para a descricao e saida do arquivo as definicoes do Manual do Usuario do
Modelo de Planejamento da Operacao de Sistemas Hidrotermicos Interligados de Longo e Medio Prazos do Projeto NEWAVE versao 27 de dezembro/2019 - pagina 158.
}
\examples{
\dontrun{
leituraNwlistcfCortes("C:/PDE2027_Caso080/nwlistcf.rel")
}

}
