% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraNwlistopGenericaComPatamaresTotal.R
\name{leituraNwlistopGenericaComPatamaresTotal}
\alias{leituraNwlistopGenericaComPatamaresTotal}
\title{Leitor de tabela Nwlistop COM patamar}
\usage{
leituraNwlistopGenericaComPatamaresTotal(
  pasta,
  nomeTabela,
  passo,
  colunaInicialJaneiro,
  pat = FALSE
)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com as tabelas do Nwlistop}

\item{nomeTabela}{nome da tabela do Nwlistop ex (efiol)}

\item{passo}{tamanho do campo}

\item{colunaInicialJaneiro}{coluna em que inicia a impressao dos dados (coluna posterior ao final da impressao do patamar)}

\item{pat}{TRUE retorna os dados por patamar, FALSE retorna por total}
}
\value{
\code{df.dadosNwlistop} data frame com os dados lidos
\itemize{
\item codigo do REE (\code{$codREE})
\item serie (\code{$serie})
\item patamar (\code{$patamar})
\item valor de ano e mes (\code{$anoMes})
\item valor dos dados da tabela escolhida (\code{$dados})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com informacao de inicio e fim da coluna 
Nao retorna os valores de media, desvio e etc. do arquivo de origem. 
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. 
Esse "problema" acontece na numeracao das series historicas. 
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraNwlistopGenericaComPatamaresTotal("C:/PDE2027_Caso080", "ghtot", 9, 12)}

}
