% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraPARP.R
\name{leituraPARP}
\alias{leituraPARP}
\title{Leitor dos dados do arquivo PARP}
\usage{
leituraPARP(pasta)
}
\arguments{
\item{pasta}{localizacao do arquivo do NEWAVE PARP}
}
\value{
\code{lt.dadosParp} lista com data frames com dados das usinas hidroeletricas para cada reservatorio equivalente (REE)
\itemize{
\item \code{df.energiahist} data frame com a energia natural afluente historica (ENA)
\itemize{
\item nome da usina (\code{$nomeUsina})
\item configuracao (\code{$configuracao})
\item ano e mes da ENA para uma determinada configuracao (\code{$anomes})
\item energia natural afluente historica (ENA) (\code{$enahist})
}
\item \code{df.parametros} data frame com dados de evaporacao mensal por usina
\itemize{
\item nome da usina (\code{$nomeUsina})
\item ano e mes da ENA para uma determinada configuracao (\code{$anomes})
\item ordem do modelo PARP antes do processo de reducao da ordem (\code{$ordemOriginal})
\item ordem do modelo PARP apos o processo de reducao da ordem (\code{$ordemFinal})
\item parametro fi do modelo PARP, lag1 (\code{$fiLag1})
\item parametro fi do modelo PARP, lag2 (\code{$fiLag2})
\item parametro fi do modelo PARP, lag3 (\code{$fiLag3})
\item parametro fi do modelo PARP, lag4 (\code{$fiLag4})
\item parametro fi do modelo PARP, lag5 (\code{$fiLag5})
\item parametro fi do modelo PARP, lag6 (\code{$fiLag6})
\item parametro coef (fi*desvio padrao t / desvio padrao t-1) do modelo PARP, lag1 (\code{$coefLag1})
\item parametro coef (fi*desvio padrao t / desvio padrao t-1) do modelo PARP, lag2 (\code{$coefLag2})
\item parametro coef (fi*desvio padrao t / desvio padrao t-1) do modelo PARP, lag3 (\code{$coefLag3})
\item parametro coef (fi*desvio padrao t / desvio padrao t-1) do modelo PARP, lag4 (\code{$coefLag4})
\item parametro coef (fi*desvio padrao t / desvio padrao t-1) do modelo PARP, lag5 (\code{$coefLag5})
\item parametro coef (fi*desvio padrao t / desvio padrao t-1) do modelo PARP, lag6 (\code{$coefLag6})
\item parametro fiA do modelo PARP com parcela anual, lag1 (\code{$fiLag1})
\item parametro fiA do modelo PARP com parcela anual, lag2 (\code{$fiLag2})
\item parametro fiA do modelo PARP com parcela anual, lag3 (\code{$fiLag3})
\item parametro fiA do modelo PARP com parcela anual, lag4 (\code{$fiLag4})
\item parametro fiA do modelo PARP com parcela anual, lag5 (\code{$fiLag5})
\item parametro fiA do modelo PARP com parcela anual, lag6 (\code{$fiLag6})
\item parametro coefA (fi*desvio padrao t / desvio padrao t-1) do modelo PARP com parcela anual, lag1 (\code{$coefLag1})
\item parametro coefA (fi*desvio padrao t / desvio padrao t-1) do modelo PARP com parcela anual, lag2 (\code{$coefLag2})
\item parametro coefA (fi*desvio padrao t / desvio padrao t-1) do modelo PARP com parcela anual, lag3 (\code{$coefLag3})
\item parametro coefA (fi*desvio padrao t / desvio padrao t-1) do modelo PARP com parcela anual, lag4 (\code{$coefLag4})
\item parametro coefA (fi*desvio padrao t / desvio padrao t-1) do modelo PARP com parcela anual, lag5 (\code{$coefLag5})
\item parametro coefA (fi*desvio padrao t / desvio padrao t-1) do modelo PARP com parcela anual, lag6 (\code{$coefLag6})
}
}
}
\description{
Faz a leitura do arquivo do NEWAVE com memoria de calculo do modelo autorregressivo periodico (parp.dat).
}
\examples{
\dontrun{
leituraPARP("C:/PDE2027_Caso080")
}

}
