% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraPostos.R
\name{leituraPostos}
\alias{leituraPostos}
\title{Leitor de dados de postos fluviometricos}
\usage{
leituraPostos(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{df.postos} data frame com dados de postos fluviometricos 
\itemize{
\item codigo do posto fluviometrico (\code{$posto})
\item ano inicial do historico de vazoes (\code{$anoInicial})
\item ano final do historico de vazoes (\code{$anoFinal})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados de postos fluviometricos 
(postos.dat).
Arquivo de acesso direto, nao formatado não formatado, com 320 / 600 
postos, cada registro correspondendo a um posto fluviometrico.
}
\examples{
\dontrun{
leituraPostos("C:/PDE2027_Caso080")
}

}
