% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraREE.R
\name{leituraREE}
\alias{leituraREE}
\title{Leitor dos dados dos Reservatorios Equivalentes de Energia (REE)}
\usage{
leituraREE(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{df.ree} data frame com os dados dos REE
\itemize{
  \item codigo do REE (\code{$codREE})
  \item nome (\code{$nomeREE})
  \item codigo do subsistema do REE (\code{$codSubsistema})
  }
}
\description{
Faz a leitura do arquivo do NEWAVE com informacao dos REE (ree.dat) e recupera o codigo, nome e sistema de cada REE cadastrada. Usa a funcao \code{\link{leituraArquivos}}
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraREE("C:/PDE2027_Caso080") }

}
