% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraRestricoesVazaoHidroDC.R
\name{leituraRestricoesVazaoHidroDC}
\alias{leituraRestricoesVazaoHidroDC}
\title{Leitor de dados de restricoes de vazao defluente para UHEs do DECOMP}
\usage{
leituraRestricoesVazaoHidroDC(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos DECOMP}
}
\value{
\code{df.RHQ} data frame com dados de restricoes de vazao
\itemize{
\item numero de cadastro da restricao (\code{$numRest})
\item codigo da usina no cadastro de usinas hidroeletricas (\code{$codUsina})
\item estagio do horizonte de estudo (\code{$estagio})
\item limite inferior da restricao no patamar 1 (\code{$limInf_p1})
\item limite superior da restricao no patamar 1 (\code{$limSup_p1})
\item limite inferior da restricao no patamar 2 (\code{$limInf_p2})
\item limite superior da restricao no patamar 2 (\code{$limSup_p2})
\item limite inferior da restricao no patamar 3 (\code{$limInf_p3})
\item limite superior da restricao no patamar 3 (\code{$limSup_p3})
\item coeficiente de participacao da UHE na restricao (\code{$coef})
\item tipo da restricao (\code{$tipoRest})
}
}
\description{
Faz a leitura dos dados de restricao de vazao defluente do DECOMP no arquivo (dadger.*).
Utiliza as informacoes do BLOCO 36 do DADGER
}
\examples{
\dontrun{
leituraRestricoesVolumeHidroDC("C:/Pasta_Deck_DECOMP")
}

}
