% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraRestricoesVolumeHidroDC.R
\name{leituraRestricoesVolumeHidroDC}
\alias{leituraRestricoesVolumeHidroDC}
\title{Leitor de dados de restricoes de volume armazenado para UHEs do DECOMP}
\usage{
leituraRestricoesVolumeHidroDC(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos DECOMP}
}
\value{
\code{df.RHV} data frame com dados de restricoes de volume armazenado
\itemize{
\item numero de cadastro da restricao (\code{$numRest})
\item codigo da usina no cadastro de usinas hidroeletricas (\code{$codUsina})
\item estagio do horizonte de estudo (\code{$estagio})
\item limite inferior da restricao (\code{$limInf})
\item limite superior da restricao (\code{$limSup})
\item coeficiente de participacao da UHE na restricao (\code{$coef})
\item tipo da restricao (\code{$tipoRest})
}
}
\description{
Faz a leitura dos dados de restricao de volume armazenado do DECOMP no arquivo (dadger.*).
Utiliza as informacoes do BLOCO 35 do DADGER
}
\examples{
\dontrun{
leituraRestricoesVolumeHidroDC("C:/Pasta_Deck_DECOMP")
}

}
