% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraSubmotorizacaoREE.R
\name{leituraSubmotorizacaoREE}
\alias{leituraSubmotorizacaoREE}
\title{Leitor dos dados de submotorizacao por REE}
\usage{
leituraSubmotorizacaoREE(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{df.submotorizacao} data frame com os valores de submotorizacao 
por REE
\itemize{
\item nome do REE (\code{$nomeREE})
\item valor de ano e mes (\code{$anoMes})
\item energia de submotorizacao [MWmes] (\code{$submotorizacao})
}
}
\description{
Faz a leitura do arquivo do relatorio de saida do NEWAVE que contem a 
informacao de submotorizacao por REE. 
Usa a funcao \code{\link{leituraArquivos}}
}
\examples{
\dontrun{
leituraSubmotorizacaoREE("C:/PDE2027_Caso080")
}

}
