% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraUsinasNaoSimuladas.R
\name{leituraUsinasNaoSimuladas}
\alias{leituraUsinasNaoSimuladas}
\title{Leitor dos dados das usinas nao simuladas por subsistema/submercado}
\usage{
leituraUsinasNaoSimuladas(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{df.UsinasNaoSimuladas} data frame com os dados das usinas nao simuladas
\itemize{
\item numero do subsistema/submercado (\code{$codSubsistema})
\item numero do bloco de usinas nao simuladas (\code{$codBlocoUNS})
\item nome do bloco de usinas nao simuladas (\code{$nomeBlocoUNS})
\item ano e mes do dado (\code{$anoMes})
\item energia em MWmedios (\code{$geracaoUNS})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados das usinas nao simuladas por subsistema/submercado (sistema.*)
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraUsinasNaoSimuladas("C:/PDE2027_Caso080")
}

}
