% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraValorDagua.R
\name{leituraValorDagua}
\alias{leituraValorDagua}
\title{Leitor de dados do valor dagua por REE}
\usage{
leituraValorDagua(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados do valor dagua}
}
\value{
\code{df.valorDagua} data frame com os valores dagua
\itemize{
\item codigo do REE (\code{$codREE})
\item serie (\code{$serie})
\item valor de ano e mes (\code{$anoMes})
\item valor dagua [$/MWh] (\code{$valorDagua})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados do valor dagua por REE  (vaguaxxx.*) e recupera esses valores por ano, mes e serie.
Nao retorna os valores de media, desvio e etc. do arquivo de origem.
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas.
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraValorDagua("C:/PDE2027_Caso080")
}

}
