% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraVazoes.R
\name{leituraVazoes}
\alias{leituraVazoes}
\title{Leitor de dados de vazoes historicas}
\usage{
leituraVazoes(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{df.vazoes} data frame com dados de vazoes naturais afluentes 
as usinas hidroeletricas
\itemize{
\item codigo do posto fluviometrico (\code{$posto})
\item ano do historico de vazoes (\code{$ano})
\item mes do historico de vazoes (\code{$mes})
\item vazao natural afluente (\code{$vazao})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com dados de vazoes naturais historicas 
afluentes as usinas hidroeletricas (vazoes.dat).
Arquivo de acesso direto, nao formatado não formatado, com 320 / 600 
postos, cada registro correspondendo a um mes do historico.
}
\examples{
\dontrun{
leituraVazoes("C:/PDE2027_Caso080")
}

}
