% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraVertimentoFioDaguaTurbinavelSubmercado.R
\name{leituraVertimentoFioDaguaTurbinavelSubmercado}
\alias{leituraVertimentoFioDaguaTurbinavelSubmercado}
\title{Leitor de dados de vertimento fio dagua turbinavel por submercado}
\usage{
leituraVertimentoFioDaguaTurbinavelSubmercado(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de vertimento fio dagua turbinavel}
}
\value{
\code{df.vertimentoFioDaguaTurbinavelSubmercado} data frame com os valores de vertimento fio dagua turbinavel por submercado
\itemize{
\item codigo do submercado (\code{$codSubmercado})
\item serie (\code{$serie})
\item valor de ano e mes (\code{$anoMes})
\item valor de vertimento [MWmes] (\code{$vertimento})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de vertimento fio dagua turbinavel por submercado (verturbmxxx.*) e recupera esses valores por ano, mes e serie.
Nao retorna os valores de media, desvio e etc. do arquivo de origem. 
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas. 
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraVertimentoFioDaguaTurbinavelSubmercado("C:/PDE2027_Caso080")}

}
