% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraVolArmUsina.R
\name{leituraVolArmUsina}
\alias{leituraVolArmUsina}
\title{Leitor de dados de volume armazenado por usina}
\usage{
leituraVolArmUsina(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE com dados de volume armazenado por usina}
}
\value{
\code{df.volArmUsina} data frame com os valores de volume armazenado por usina
\itemize{
\item codigo da usina (\code{$codUsina})
\item serie (\code{$serie})
\item valor de ano e mes (\code{$anoMes})
\item valor do volume armazenado [hm3] (\code{$volume})
}
}
\description{
Faz a leitura dos arquivos do NEWAVE com dados de volume armazenado por usina  (varmuhxxx.*) e 
recupera esses valores por codigo da usina, ano, mes e serie.
Nao retorna os valores de total, media, desvio e etc. do arquivo de origem. 
Faz uma modificacao no numero da serie para garantir compatibilidade da sequencia. Esse "problema" acontece na numeracao das series historicas. 
Assim troca-se o valor original para o campo serie (ano) pelo valor dentro de uma mesma sequencia para cada ano.
}
\examples{
\dontrun{
leituraVolArmUsina("C:/PDE2027_Caso080")}

}
