% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculoMediaDpCmo.R
\name{calculoMediaDpCMO}
\alias{calculoMediaDpCMO}
\title{Calculo da media e desvio padrao do CMO}
\usage{
calculoMediaDpCMO(pasta)
}
\arguments{
\item{pasta}{localizacao dos arquivos do NEWAVE para os quais deseja efetuar o calculo}
}
\value{
\code{df.cmoMedio} data frame com os valores de custo marginal de demanda
\itemize{
\item codigo do submercado (\code{$codSubmercado})
\item valor do ano (\code{$ano})
\item media do CMO anual [$/MWh] (\code{$mediaCMO})
\item desvio padrao da media do CMO [$/MWh] (\code{$dpCMO})
}
}
\description{
Faz o calculo da media e do desvio padrao dos CMOs por submercado e por ano de estudo,
reproduzindo o calculo que e encontrado ao final do relatorio de saida do NEWAVE
}
\examples{
\dontrun{
calculoMediaDpCMO("C:/PDE2030_Caso080")}

}
