% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leituraSubmotorizacaoREE.R
\name{leituraSubmotorizacaoREE}
\alias{leituraSubmotorizacaoREE}
\title{Leitor dos dados de submotorizacao por REE}
\usage{
leituraSubmotorizacaoREE(pastaCaso)
}
\arguments{
\item{pastaCaso}{caracter com localizacao dos arquivos NEWAVE}
}
\value{
\code{df.submotorizacao} data frame com os valores de submotorizacao por REE
\itemize{
\item nome do REE (\code{$nomeREE})
\item valor de ano e mes (\code{$anoMes})
\item energia de submotorizacao [MWmes] (\code{$submotorizacao})
}
}
\description{
Faz a leitura do arquivo do NEWAVE com informacao de submotorizacao por REE (relat.d**). Usa a funcao \code{\link{leituraArquivos}}
Usa como referencia para a leitura do arquivo as posicoes definidas no Manual do Usuario do
Modelo Estrategico de geracao hidrotermica a subsistemas equivalentes do Projeto NEWAVE
}
\examples{
\dontrun{
leituraSubmotorizacaoREE("C:/PDE2027_Caso080")}

}
